Edit API to allow better translations
Locale files have not been updated yet. This basically provide a way to
add more specific translations when required.

This commit does not change translations display because this would
implies around 90×60 translation to create in each language and I do not
know them enought to provide translations. Also if we enable custom
translation (from .tr files) on a description of a node, this
description would default to english for missing translations, while for
now this provides imperfects translation (uppercase issue + does not
allow to have `Outer Jungle Wood Stair` _and_ a correct translation, like in `stairs` mod (we are limited to `Jungle Wood Outer
Stair` unless we provide specific translations)).